+++ shairport.c
@@ -42,8 +42,8 @@
 
 #include "config.h"
 
-#ifdef HAVE_LIBPOLARSSL
-#include <polarssl/md5.h>
+#ifdef HAVE_LIBMBEDTLS
+#include <mbedtls/md5.h>
 #endif
 
 #ifdef HAVE_LIBSSL
@@ -109,8 +109,8 @@ char* get_version_string() {
   char* version_string = malloc(200);
   if (version_string) {
     strcpy(version_string, PACKAGE_VERSION);
-  #ifdef HAVE_LIBPOLARSSL
-    strcat(version_string, "-PolarSSL");
+  #ifdef HAVE_LIBMBEDTLS
+    strcat(version_string, "-mbedTLS");
   #endif
   #ifdef HAVE_LIBSSL
     strcat(version_string, "-OpenSSL");
@@ -1046,11 +1046,11 @@ int main(int argc, char **argv) {
   MD5_Final(ap_md5, &ctx);
 #endif
 
-#ifdef HAVE_LIBPOLARSSL
-  md5_context tctx;
-  md5_starts(&tctx);
-  md5_update(&tctx, (unsigned char *)config.service_name, strlen(config.service_name));
-  md5_finish(&tctx, ap_md5);
+#ifdef HAVE_LIBMBEDTLS
+  mbedtls_md5_context tctx;
+  mbedtls_md5_starts(&tctx);
+  mbedtls_md5_update(&tctx, (unsigned char *)config.service_name, strlen(config.service_name));
+  mbedtls_md5_finish(&tctx, ap_md5);
 #endif
   memcpy(config.hw_addr, ap_md5, sizeof(config.hw_addr));
 #ifdef CONFIG_METADATA